requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.11
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fonttools==4.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.24
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.38
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest_asyncio==1.6.0
numpy==2.2.5
openai==1.76.0
orjson==3.10.16
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack_data==0.6.3
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
zstandard==0.23.0
